backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.2
pydantic-settings==2.2.1
requests==2.32.2
httpx>=0.27.0
apscheduler>=3.10.0
beautifulsoup4==4.12.3
python-dotenv>=1.0.0
